requirements.txt
numpy
python-telegram-bot
requests
requests-cache
